refactored with SpatialInstrumentInternal and more accessible SpatialInstrument
